pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ex-ai-mcp-server"
version = "2.0.0"
description = "Production-ready EX-AI MCP Server with intelligent routing"
authors = [{name = "Zazzles", email = "zazzles@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "openai>=1.55.2",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # DEPENDENCY FIX (2025-11-05): Switched from zhipuai to zai-sdk
    # zhipuai requires PyJWT<2.9.0, conflicting with MCP 1.20.0 (requires PyJWT>=2.10.1)
    # zai-sdk is compatible and provides better features
    "zai-sdk>=0.0.3.3",
    "httpx>=0.28.0",
    "importlib-resources>=5.0.0; python_version<'3.9'",
]

[project.urls]
Homepage = "https://github.com/Zazzles2908/EX-AI-MCP-Server"
Issues = "https://github.com/Zazzles2908/EX-AI-MCP-Server/issues"
Repository = "https://github.com/Zazzles2908/EX-AI-MCP-Server"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

remote = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.0.0",
    "fastmcp>=1.0.0",
]

[project.scripts]
ex-ai-mcp-server = "server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
